requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
pandas
langchain
langchain_community
tqdm
langchain_groq
pyodbc
jq
faiss-cpu
openpyxl
matplotlib
transformers>=4.31.0
sentence-transformers>=2.2.2
torch>=1.12.1
bs4==0.0.2
html2text==2020.1.16
chromadb==0.4.22
streamlit
langchain_core
scikit-learn
python-dotenv
ipykernel
streamlit-extras

